INC-126 feat: add CommonTableWithPagination and move folder commonTable into commonTableWithPagination
