+++ src/components/ui/commonTableWithPagination/commonTableWithPagination.module.scss
@@ -1,0 +1,7 @@
+.pagination {
+  margin-top: 36px;
+}
+
+.text {
+  text-align: center;
+}
