+++ src/components/ui/commonTableWithPagination/index.ts
@@ -1,0 +1,2 @@
+export * from './CommonTableWithPagination'
+export * from './commonTable'
